bug fixes
Fixes a merge issue and to make sure N and NY are correctly calculated and there is a full record for each water year. These bugs were caused by the end of the timeseries having NAs that were not correctly removed.

Added m-1 days to end of timeseries to get the rolling average all the way to the end of the water year/seasonal period.

Find NA days based on rolling average flow instead of daily flow. This will correctly recognize they are part of the next water year with an imcomplete record and remove them.